Walk!
Making it a bit more interesting--

You're prompted for input, and that's chopped up into an array of words.
The first word is the command, and everything else is an argument to that
command. The hash `%command-to-method-map` maps out which commands are
available.

Interesting bits of syntax:

Because `walk` is equivalent to `walk()` (that is, parentheses are optional),
you have to use `&walk` to pass around the sub itself (and avoid invoking
`walk`) (similarish to passing around methods in Ruby).

The slurpy/splatty operators are not symmetrical like in some languages.
`*` to slurp something up, and `|` to ... puke it out? What's the opposite
of slurp? lol