Add a 2d map system with example map.
Some pretty substantial changes 😛.

There's now a map/map-making system to use.

It's a very simple abstraction at the moment (and probably kinda bad):
A `SimpleMap` is just a multi-dimensional array with a few methods to
help you ensure that your current index positions are valid.

```
[[...PPP.],
 [.PPP.P.],
 [..P..P.],
 [..PPPPP],
 [......P]]
```

(imagine each P is a `Place` object and each dot is `Any`)

`Place` is a similarly simple abstraction --pretty much just a struct
for a name and description (although it has a `seen` attribute which
nicely allows us to show the `description` just once, and `name` any
other visit).

Enclosed my code in one giant class as I really appreciate "hoisting"
for readability (being able to define `main` up-front and just not
having to worry too much about the order of my definitions).